fix(pipeline): 4 compile errors from run 34929267825
CI-log-verified (win job 104253863137):
- E0432 trace.rs:12 — BackgroundModel is an isg-native pipeline type
  (crate::pipeline::background), not isg-core; fix the import path
- E0308 quantize.rs:226-230 — centroid update: clamp-cast the i64 mean
  back to i32 before unwrap_or/comparison/assignment

+++ crates/isg-native/src/pipeline/quantize.rs
@@ -224,7 +224,7 @@ fn kmeans(samples: &[[i32; 3]], k: u32, iters: u32) -> (Vec<[i32; 3]>, Vec<u8>) 
                 let old = centres[ci][ch];
                 let next = sums[ch]
                     .checked_div(cnt as i64)
-                    .map(|v| v.clamp(0, 255))
+                    .map(|v| v.clamp(0, 255) as i32)
                     .unwrap_or(old);
                 if next != old {
                     centres[ci][ch] = next;
